restructured kratos IO in module
