+++ stem/mesh.py
@@ -2,9 +2,8 @@ from typing import Dict, List, Tuple, Union, Any
 
 import numpy as np
 import numpy.typing as npt
-#from gmsh_utils.gmsh_IO import GmshIO
 
-from stem.kratos_IO import KratosIO
+from stem.IO.kratos_IO import KratosIO
 
 class Node:
     """
